# proc: do not load g0 until it's needed when stacktracing